requirements.txt
absl-py==2.1.0
Bottleneck @ file:///work/perseverance-python-buildout/croot/bottleneck_1698877801136/work
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.0
idna==3.6
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numexpr @ file:///work/perseverance-python-buildout/croot/numexpr_1698871031164/work
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp312-cp312-linux_x86_64.whl#sha256=71892d12f82a9c47262bffc99a1edd8ebc0b3d1e033366094bd45a8b4c7c2d43
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.8
packaging==23.2
pandas @ file:///croot/pandas_1702317985682/work/dist/pandas-2.1.4-cp312-cp312-linux_x86_64.whl#sha256=e08f6f54958b7b5f7624ec6b206777bd22f94786e77f06f8d346e758d3da3fac
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///work/perseverance-python-buildout/croot/pytz_1698849876288/work
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
setuptools==68.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.9.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==2.2.0
Werkzeug==3.0.1
wheel==0.41.2